Add correct icon
